+++ Changelog.md
@@ -1,3 +1,6 @@
+## 1.7
+- Fix AntiBlink issues where it was comparing objects from different landblocks
+
 ## 1.6
 - Skip AntiCheat checks for admins / cloaked players (see settings to disable)